requirements.txt
pandas
numpy
netCDF4
cftime
notebook
pyyaml
